smol_gesture_package/requirements.txt
torch
transformers
Pillow
opencv-python
